Code refactor to dojot package
