+++ pom.xml
@@ -1,6 +1,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>br.com.cpqd.mutualauthentication.registry</groupId>
+  <groupId>br.com.dojot.mutualauthentication.registry</groupId>
   <artifactId>Registry</artifactId>
   <version>1.1.0-SNAPSHOT</version>
   
@@ -18,7 +18,7 @@
 		</dependency>
 		<!-- Communication Microservice -->
 		<dependency>
-			<groupId>br.com.cpqd.mutualauthentication.communication</groupId>
+			<groupId>br.com.dojot.mutualauthentication.communication</groupId>
 			<artifactId>Communication-lib</artifactId>
 			<version>${communication.version}</version>
 		</dependency>
